update interactive Framework

Former-commit-id: 0c9887b1d9d43b5b4613c5799e46050b1807d132

+++ PythonVersion/com/cleanEnv.py
@@ -6,30 +6,47 @@ import getopt
 from config.config import Config
 from config.config import ConfigKeys as Keys
 import subprocess
+from sys import platform
 
 class CleanEnv(object):
     @staticmethod
     def kill_karmaRun():
         kill_karmaRun=subprocess.Popen(' ps aux | grep \'karma run\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1 ', shell=True)
-        kill_karmaRun.wait()
+        kill_karmaRun.communicate()
     @staticmethod
     def kill_karmaStart():
-    	kill_karmaStart=subprocess.Popen('ps aux | grep \'karma start\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
-        kill_karmaStart.wait()
-        kill_karmaStart=subprocess.Popen('ps aux | grep chrome | grep -v \'grep\' | awk \'{print $2}\' |xargs kill -9 >/dev/null 2>&1', shell=True)
-        kill_karmaStart.wait()
+        kill_karmaStart=subprocess.Popen('ps aux | grep \'karma start\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
+        kill_karmaStart.communicate()
+    @staticmethod
+    def kill_karmaPort():
+        kill_port=subprocess.Popen('lsof -i:9877 | grep \'*:9877\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
+        kill_port.communicate()
+        kill_port=subprocess.Popen('lsof -i:9878 | grep \'*:9878\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
+        kill_port.communicate()
+    @staticmethod
+    def kill_Chrome():
+        if platform == "darwin":
+            kill_karmaStart=subprocess.Popen('killall -9 \"Google Chrome\"', shell=True)
+            kill_karmaStart.communicate()
+        if platform == "linux" or platform == "linux2":
+            kill_karmaStart=subprocess.Popen('ps aux | grep chrome | grep -v \'grep\' | awk \'{print $2}\' |xargs kill -9 >/dev/null 2>&1', shell=True)
+            kill_karmaStart.communicate()
+    @staticmethod
+    def kill_Safari():
+        kill_karmaStart=subprocess.Popen('ps aux | grep Safari | grep -v \'grep\' | awk \'{print $2}\' |xargs kill -9 >/dev/null 2>&1', shell=True)
+        kill_karmaStart.communicate()
     @staticmethod
     def kill_Firefox():
         kill_karmaStart=subprocess.Popen('ps aux | grep firefox | grep -v \'grep\' | awk \'{print $2}\' |xargs kill -9 >/dev/null 2>&1', shell=True)
-        kill_karmaStart.wait()
+        kill_karmaStart.communicate()
     @staticmethod
     def kill_lockServer():
         kill_karmaStart=subprocess.Popen('ps aux | grep \'lockserver.jar\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
-        kill_karmaStart.wait()
+        kill_karmaStart.communicate()
     @staticmethod
     def kill_runTest():
         kill_runtest=subprocess.Popen('ps aux | grep \'runTest.py\' | grep -v \'grep\' | awk \'{print $2}\'|xargs kill -9 >/dev/null 2>&1', shell=True)
-        kill_runtest.wait()
+        kill_runtest.communicate()
     @staticmethod
     def rm_JsTestResult(testResultFile,mode):
         if mode == "P2P":
@@ -45,5 +62,3 @@ class CleanEnv(object):
     @staticmethod
     def rm_TestResult(testresultFile):
         RenameCase=subprocess.Popen("rm "+ testresultFile, shell=True)
-
-
